Use more floats (save 1k of Flash)

+++ src/math_osm.cpp
@@ -7,12 +7,12 @@
 // we return float here, because we need the fraction
 
 float lon2tilex(float lon, uint8_t z) {
-    return (lon + 180.0) / 360.0 * (float)(1 << z);
+    return (lon + 180.0f) / 360.0f * (float)(1 << z);
 }
 
 float lat2tiley(float lat, uint8_t z) {
     float latrad = lat * PI / 180.0f;
-    return (1.0 - asinh(tanf(latrad)) / PI) / 2.0f * (float)(1 << z);
+    return (1.0f - asinh(tanf(latrad)) / PI) / 2.0f * (float)(1 << z);
 }
 
 // helper function to get the offset within the tile
@@ -30,10 +30,10 @@ float tiley2lat(float y, uint8_t z) {
     return 180.0f / PI * atanf(0.5f * (expf(n) - expf(-n)));
 }
 
-float osmResolution[] = {156543.03, 78271.52, 39135.76, 19567.88, 9783.94, 4891.97, 2445.98, 1222.99, 611.50, 305.75,
-                         152.87,    76.43,    38.21,    19.10,    9.55,    4.77,    2.38,    1.19,    0.59
+float osmResolution[] = {156543.03f, 78271.52f, 39135.76f, 19567.88f, 9783.94f, 4891.97f, 2445.98f, 1222.99f, 611.50f, 305.75f,
+                         152.87f,    76.43f,    38.21f,    19.10f,    9.55f,    4.77f,    2.38f,    1.19f,    0.59f
                         };
 
 float getTileResolution(float lat, uint8_t z) {
-    return 156543.03 /*meters/pixel*/ * cos(lat) / (2 ^ z);
+    return 156543.03f /*meters/pixel*/ * cosf(lat) / (2 ^ z);
 }
